requirements.txt
streamlit>=1.32.0
pybaseball>=2.2.7
MLB-StatsAPI>=1.7.2
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
joblib>=1.3.0
requests>=2.31.0
